requirements.txt
anyio==4.4.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
Faker==25.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-telegram-bot==13.6
pytz==2024.1
requests==2.26.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
tornado==6.1
tzlocal==5.2
urllib3==1.26.7
wheel==0.43.0
